+++ src/main/java/com/be08/smart_notes/repository/DocumentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.be08.smart_notes.entity.DocumentEntity;
 
-public interface DocumentRepository extends JpaRepository<DocumentEntity, Long>{
+public interface DocumentRepository extends JpaRepository<DocumentEntity, Integer>{
 	
 }
